+++ kfinance/meta_classes.py
@@ -7,13 +7,19 @@ from cachetools import LRUCache, cached
 import numpy as np
 import pandas as pd
 
-from .constants import LINE_ITEMS, BusinessRelationshipType, PeriodType, SegmentType
+from .constants import (
+    LINE_ITEMS,
+    BusinessRelationshipType,
+    CompetitorSource,
+    PeriodType,
+    SegmentType,
+)
 from .fetch import KFinanceApiClient
 from .pydantic_models import RelationshipResponse
 
 
 if TYPE_CHECKING:
-    from .kfinance import BusinessRelationships
+    from .kfinance import BusinessRelationships, Companies
 
 logger = logging.getLogger(__name__)
 
@@ -433,25 +439,42 @@ class CompanyFunctionsMetaClass:
         if not self._company_descriptions:
             self._company_descriptions = self.kfinance_api_client.fetch_company_descriptions(company_id=self.company_id)
         return self._company_descriptions["description"]
-    
+
     @property
     def alternate_names(self):
         if not self._company_other_names:
             self._company_other_names = self.kfinance_api_client.fetch_company_other_names(company_id=self.company_id)
         return self._company_other_names["alternate_names"]
-    
+
     @property
     def historical_names(self):
         if not self._company_other_names:
             self._company_other_names = self.kfinance_api_client.fetch_company_other_names(company_id=self.company_id)
         return self._company_other_names["historical_names"]
-    
 
     @property
     def native_names(self):
         if not self._company_other_names:
             self._company_other_names = self.kfinance_api_client.fetch_company_other_names(company_id=self.company_id)
         return self._company_other_names["native_names"]
+
+    def competitors(
+        self, competitor_source: CompetitorSource = CompetitorSource.all
+    ) -> "Companies":
+        """Get the list of companies that are competitors of company_id, optionally filtered by the competitor_source type.
+
+        :return: The list of companies that are competitors of company_id, optionally filtered by the competitor_source type
+        :rtype: Companies
+        """
+        from .kfinance import Companies
+
+        competitors_data = self.kfinance_api_client.fetch_competitors(
+            company_id=self.company_id, competitor_source=competitor_source
+        )["companies"]
+        return Companies(
+            kfinance_api_client=self.kfinance_api_client,
+            company_ids=[company["company_id"] for company in competitors_data],
+        )
 
 
 for line_item in LINE_ITEMS: