+++ kfinance/CHANGELOG.md
@@ -1,8 +1,13 @@
 # Changelog
 
+## v2.6.0
+ - Add mergers & acquisitions.
+
+## v2.5.0
+ - Add competitors function and llm tools
+
 ## v2.4.4
 - Replace `cached_property` with `property` to enable batching
-
 
 ## v2.4.3
 - Improve MCP tool build process